Archives/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-camera-imagefield==0.2
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.4.0
python-datauri==1.1.0
reportlab==3.6.12
sqlparse==0.4.3
